openEuler: arm64/mpam: Probe partid,pmg and feature capabilities' ranges from classes
to #34407882

commit d63bf6e6ec573a1b0e4115efcc818f7922ff2ac5 openEuler

hulk inclusion
category: feature
feature: ARM MPAM support
bugzilla: 48265
CVE: NA

--------------------------------

So far partid and pmg has been probed by mpam classes, so do feature
capabilities of each resources, for resctrl intermediate processing
layer, those information placed in classes should be restored in
internal resctrl resource structure.

For simplicity, capabilities related are unifiedly controlled by
integer input, also should its' width probed from mpam classes.
currently we only give priority width and hardlimit width fields
in resctrl resource structure, in order to adapt more features,
part of this would be re-covered.

+++ arch/arm64/kernel/mpam/mpam_resctrl.c
@@ -308,7 +308,7 @@ void closid_init(void)
 	for_each_resctrl_resource(r) {
 		if (r->alloc_enabled) {
 			rr = r->res;
-			num_closid = min(num_closid, rr->num_partid);
+			num_closid = min(num_closid, (int)rr->num_partid);
 		}
 	}
 	closid_free_map = BIT_MASK(num_closid) - 1;
